+++ test/TestAsset/WithMethodsNamedAfterKeywords/WithoutReturnTypeDeclaration.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * @see       https://github.com/zendframework/zend-file for the canonical source repository
+ * @copyright Copyright (c) 2018 Zend Technologies USA Inc. (https://www.zend.com)
+ * @license   https://github.com/zendframework/zend-file/blob/master/LICENSE.md New BSD License
+ */
+
+namespace ZendTest\File\TestAsset\WithMethodsNamedAfterKeywords;
+
+final class WithoutReturnTypeDeclaration
+{
+    public function class()
+    {
+    }
+
+    public function interface()
+    {
+    }
+
+    public function trait()
+    {
+    }
+}